:racehorse: Remove type-unstable literals in generic code
Several helper functions with a generic element type `T` used `Float64`
literals in clamps and normalizations, so their documented `NTuple{6,
T}` and `SVector{6, T}` return contracts were violated for any element
type other than `Float64`. Replace the literals with `T`-typed values
in the element normalization, the mean and osculating conversions, the
dynamics clamps, and the J2 squared regularization. Also clamp the
altitude with `zero(h)` in the drag acceleration and force a
homogeneous `Float64` time span even when the user passes an integer
`tf`.

The results are bit-identical for `Float64` inputs.

+++ ext/SatelliteAnalysisDecayExt/main.jl
@@ -104,7 +104,8 @@ function _decay_analysis(
     ā, ē, ī, Ω̄, ω̄, M̄ = _osculating_to_mean_elements(orb.a, orb.e, orb.i, orb.Ω, orb.ω, M)
     u = Vector(_classical_to_equinoctial(ā, ē, ī, Ω̄, ω̄, M̄))
 
-    tspan = (0.0, tf)
+    # Force a homogeneous time span even if the user passes an integer `tf`.
+    tspan = (0.0, Float64(tf))
 
     # Pre-allocate the J2 osculating propagator used for the mean-to-osculating
     # conversion inside the drag quadrature, avoiding one propagator allocation per
